Move app logic into a separate file
